+++ tests/testthat/test-ui-param.R
@@ -183,4 +183,35 @@ if (!.Call(`_rxode2_isIntel`)) {
                                     state = character(0))))
 
   })
+
+  test_that("linCmt single compartment parses correctly", {
+
+    mod <- function() {
+      ini({
+        cl <- 0.1
+        vc <- 4
+        err.sd <- 0.1
+      })
+      model({
+        Cc <- linCmt(cl, vc)
+        Cc ~ add(err.sd)
+      })
+    }
+
+    mod <- mod()
+
+    expect_error(mod$params, NA)
+
+    expect_equal(mod$params,
+                 list(pop = c("cl", "vc"),
+                      resid = "err.sd",
+                      group = structure(list(), names = character(0)),
+                      linCmt = TRUE,
+                      cmt = "central",
+                      output = list(primary = character(0),
+                                    secondary = character(0),
+                                    endpoint = "Cc",
+                                    state = character(0))))
+
+  })
 }